save work

+++ part3/hbnb/app/api/v1/reviews.py
@@ -17,9 +17,26 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @api.response(400, 'You cannot review your own place')
+    @api.response(400, 'You have already reviewed this place')
+    @jwt_required()
     def post(self):
         """Register a new review"""
+        current_user = get_jwt_identity()
         review_data = api.payload
+
+        review_data['user_id'] = current_user
+
+        place = facade.get_place(review_data['place_id'])
+        if not place:
+            return {'error': 'Place not found'}, 400
+        
+        if place.owner_id == current_user:
+            return {'error': 'You cannot review your own place'}, 400
+        
+        review_exists = facade.get_review_user_place(current_user, review_data['place_id'])
+        if review_exists:
+            return {'error': 'You have already reviewed this place'}
 
         try:
             new_review = facade.create_review(review_data)
@@ -48,28 +65,45 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @api.response(403, 'Unauthorized action')
+    @jwt_required()
     def put(self, review_id):
         """Update a review's information"""
+        current_user = get_jwt_identity()
         review_data = api.payload
 
-        update_review = facade.update_review(review_id, review_data)
-        if not update_review:
+        review = facade.get_review(review_id)
+
+        if not review:
             return {'error': 'Review not found'}, 404
-        
+
+        if review['user_id'] != current_user:
+            return {'error': 'Unauthorized action'}, 403
+
         try:
-            facade.update_review(review_id, review_data)
-            return update_review.to_dict(), 200
+            review_data['user_id'] = current_user
+
+            updated_review = facade.update_review(review_id, review_data)
+            return updated_review.to_dict(), 200
+
         except Exception as error:
             return {'error': str(error)}, 400
 
     @api.response(200, 'Review deleted successfully')
+    @api.response(403, 'Unauthorized action')
     @api.response(404, 'Review not found')
+    @jwt_required()
     def delete(self, review_id):
         """Delete a review"""
+        current_user = get_jwt_identity()
         review = facade.get_review(review_id)
+
         if not review:
             return {'error': 'Review not found'}, 404
         
+        if review['user_id'] != current_user:
+            return {'error': 'Unauthorized action'}, 403
+            
         try:
             facade.delete_review(review_id)
             return {'message': 'Review deleted successfully'}, 200
